[python] Move ``AnalysisFileContext`` into its own module
Relocate ``AnalysisFileContext`` from ``eos.analysis_file_description`` into a
new leaf module ``eos.analysis_file_context``. This severs the circular import
between ``eos.analysis_file_description`` and ``eos.figure`` -- whose modules
import ``AnalysisFileContext`` -- so that ``eos.analysis_file_description`` can
import ``FigureFactory`` at the top level in a subsequent change.

Repoint all importers (``eos.figure`` modules, ``eos.tasks``, ``eos.reporting``)
to the new module, and move the corresponding test into
``analysis_file_context_TEST.py``. No behavioural change.
